Fixed issue with calculating darkProb
Changed darkProb from int to float, fixed casting issues

+++ C#/Spectroscopy Viewer/Spectroscopy Viewer/spectrum.cs
@@ -97,7 +97,7 @@ namespace Spectroscopy_Viewer
         private void createDataPlot()
         {
             int x = new int();
-            int y = new int();
+            float y = new float();
 
             for (int i = 0; i < dataSize; i++)
             {
